+++ pkg/coredata/migrations/20250501T091000Z.sql
@@ -1,0 +1,2 @@
+ALTER TABLE vendors DROP COLUMN service_start_at;
+ALTER TABLE vendors DROP COLUMN service_termination_at;