+++ pkg/coredata/vendor.go
@@ -28,28 +28,26 @@ import (
 
 type (
 	Vendor struct {
-		ID                         gid.GID    `db:"id"`
-		OrganizationID             gid.GID    `db:"organization_id"`
-		Name                       string     `db:"name"`
-		Description                *string    `db:"description"`
-		Category                   string     `db:"category"`
-		ServiceStartAt             time.Time  `db:"service_start_at"`
-		ServiceTerminationAt       *time.Time `db:"service_termination_at"`
-		HeadquarterAddress         *string    `db:"headquarter_address"`
-		LegalName                  *string    `db:"legal_name"`
-		WebsiteURL                 *string    `db:"website_url"`
-		PrivacyPolicyURL           *string    `db:"privacy_policy_url"`
-		ServiceLevelAgreementURL   *string    `db:"service_level_agreement_url"`
-		DataProcessingAgreementURL *string    `db:"data_processing_agreement_url"`
-		Certifications             []string   `db:"certifications"`
-		BusinessOwnerID            *gid.GID   `db:"business_owner_id"`
-		SecurityOwnerID            *gid.GID   `db:"security_owner_id"`
-		StatusPageURL              *string    `db:"status_page_url"`
-		TermsOfServiceURL          *string    `db:"terms_of_service_url"`
-		SecurityPageURL            *string    `db:"security_page_url"`
-		TrustPageURL               *string    `db:"trust_page_url"`
-		CreatedAt                  time.Time  `db:"created_at"`
-		UpdatedAt                  time.Time  `db:"updated_at"`
+		ID                         gid.GID   `db:"id"`
+		OrganizationID             gid.GID   `db:"organization_id"`
+		Name                       string    `db:"name"`
+		Description                *string   `db:"description"`
+		Category                   string    `db:"category"`
+		HeadquarterAddress         *string   `db:"headquarter_address"`
+		LegalName                  *string   `db:"legal_name"`
+		WebsiteURL                 *string   `db:"website_url"`
+		PrivacyPolicyURL           *string   `db:"privacy_policy_url"`
+		ServiceLevelAgreementURL   *string   `db:"service_level_agreement_url"`
+		DataProcessingAgreementURL *string   `db:"data_processing_agreement_url"`
+		Certifications             []string  `db:"certifications"`
+		BusinessOwnerID            *gid.GID  `db:"business_owner_id"`
+		SecurityOwnerID            *gid.GID  `db:"security_owner_id"`
+		StatusPageURL              *string   `db:"status_page_url"`
+		TermsOfServiceURL          *string   `db:"terms_of_service_url"`
+		SecurityPageURL            *string   `db:"security_page_url"`
+		TrustPageURL               *string   `db:"trust_page_url"`
+		CreatedAt                  time.Time `db:"created_at"`
+		UpdatedAt                  time.Time `db:"updated_at"`
 	}
 
 	Vendors []*Vendor
@@ -82,8 +80,6 @@ SELECT
     headquarter_address,
     legal_name,
     website_url,
-    service_start_at,
-    service_termination_at,
     privacy_policy_url,
     service_level_agreement_url,
     data_processing_agreement_url,
@@ -146,8 +142,6 @@ INSERT INTO
         service_level_agreement_url,
         data_processing_agreement_url,
         certifications,
-        service_start_at,
-        service_termination_at,
         business_owner_id,
         security_owner_id,
         status_page_url,
@@ -171,8 +165,6 @@ VALUES (
     @service_level_agreement_url,
     @data_processing_agreement_url,
     @certifications,
-    @service_start_at,
-    @service_termination_at,
     @business_owner_id,
     @security_owner_id,
     @status_page_url,
@@ -198,8 +190,6 @@ VALUES (
 		"service_level_agreement_url":   v.ServiceLevelAgreementURL,
 		"data_processing_agreement_url": v.DataProcessingAgreementURL,
 		"certifications":                v.Certifications,
-		"service_start_at":              v.ServiceStartAt,
-		"service_termination_at":        v.ServiceTerminationAt,
 		"business_owner_id":             v.BusinessOwnerID,
 		"security_owner_id":             v.SecurityOwnerID,
 		"status_page_url":               v.StatusPageURL,
@@ -252,8 +242,6 @@ SELECT
     service_level_agreement_url,
     data_processing_agreement_url,
     certifications,
-    service_start_at,
-    service_termination_at,
     business_owner_id,
     security_owner_id,
     status_page_url,
@@ -300,8 +288,6 @@ UPDATE vendors
 SET
 	name = @name,
 	description = @description,
-	service_start_at = @service_start_at,
-	service_termination_at = @service_termination_at,
 	category = @category,
 	headquarter_address = @headquarter_address,
 	legal_name = @legal_name,
@@ -327,8 +313,6 @@ WHERE %s
 		"updated_at":                    time.Now(),
 		"name":                          v.Name,
 		"description":                   v.Description,
-		"service_start_at":              v.ServiceStartAt,
-		"service_termination_at":        v.ServiceTerminationAt,
 		"category":                      v.Category,
 		"headquarter_address":           v.HeadquarterAddress,
 		"legal_name":                    v.LegalName,